extract roomba info to separate file
