+++ roombapy/__init__.py
@@ -1,4 +1,5 @@
 from .discovery import RoombaDiscovery
 from .getpassword import RoombaPassword
-from .roomba import Roomba, RoombaConnectionError, RoombaInfo
+from .roomba import Roomba, RoombaConnectionError
+from .roomba_info import RoombaInfo
 from .roomba_factory import RoombaFactory
